requirements.txt
csctracker-py-core==24.2.17
csctracker-queue-scheduler==24.2.1

gitpython==3.1.41
